+++ src/blocks.js
@@ -5731,7 +5731,7 @@ ReporterBlockMorph.prototype.prepareToBeGrabbed = function (handMorph) {
         this.setPosition(oldPos);
     }
     ReporterBlockMorph.uber.prepareToBeGrabbed.call(this, handMorph);
-    this.alpha = Math.min(this.alpha, 0.85);
+    handMorph.alpha = this.alpha < 1 ? 1 : 0.85;
     this.cachedSlotSpec = null;
 };